Move related files to OpenGL  directory
