+++ src/RageUtil/CMakeLists.txt
@@ -48,8 +48,8 @@ list(APPEND SMDATA_RAGE_GRAPHICS_SRC
   "Graphics/RageBitmapTexture.cpp"
   "Graphics/RageDisplay.cpp"
   "Graphics/RageDisplay_Null.cpp"
-  "Graphics/RageDisplay_OGL.cpp"
-  "Graphics/RageDisplay_OGL_Helpers.cpp"
+  "Graphics/Display/OpenGL/RageDisplay_OGL.cpp"
+  "Graphics/Display/OpenGL/RageDisplay_OGL_Helpers.cpp"
   "Graphics/RageModelGeometry.cpp"
   "Graphics/RageSurface.cpp"
   "Graphics/RageSurface_Load.cpp"
@@ -70,8 +70,8 @@ list(APPEND SMDATA_RAGE_GRAPHICS_HPP
   "Graphics/RageBitmapTexture.h"
   "Graphics/RageDisplay.h"
   "Graphics/RageDisplay_Null.h"
-  "Graphics/RageDisplay_OGL.h"
-  "Graphics/RageDisplay_OGL_Helpers.h"
+  "Graphics/Display/OpenGL/RageDisplay_OGL.h"
+  "Graphics/Display/OpenGL/RageDisplay_OGL_Helpers.h"
   "Graphics/RageModelGeometry.h"
   "Graphics/RageSurface.h"
   "Graphics/RageSurface_Load.h"
