+++ src/RageUtil/Graphics/Display/OpenGL/RageDisplay_OGL_Helpers.cpp
